// src/api/generated/schemas/travelTypeCompatibility.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */
import type { TravelTypeCompatibilityItem } from './travelTypeCompatibilityItem';

export interface TravelTypeCompatibility {
  best: TravelTypeCompatibilityItem;
  worst: TravelTypeCompatibilityItem;
}